Change default path handling for existing AD objs (#59)
Changes the behaviour of path for existing objects when no path was
specified and identity points to a valid object. The old behaviour was
to treat no path value as the default container for that object type
which would have resulted in moving the object. The new behaviour will
treat no path or empty path value as keep the existing path. The object
will only be moved if an explicit path value was set.

Adds support for the sentinel value microsoft.ad.default_path to be
specified as the path which will be the default container for that
object type.