Add dn_escape and parse_dn filters
Adds filters that can be used to escape values for use inside a DN
attribute value and to parse a DN string into a more structured object.
These filters are useful with the `microsoft.ad.ldap` inventory plugin
as well as when forming values like the `path` or other DN attributes.