Add identity server lookup base code
Add a way to specify a custom server to lookup an AD identity on when
using module options that accept a list of identities for an attribute.
For example the microsoft.ad.group members option.

Also adds the domain_credentials option that is common to add AD based
modules to specify credentials for the extra AD servers being contacted
in case the default credentials do not work.

Migrates existing options that accept a distinguishedName to the new
lookup code to align the behaviour across the modules.