microsoft.ad.user - Add group lookup DN logic
Adds the same DN lookup logic to the microsoft.ad.user groups option.
This allows the caller to add/remove/set the user's group membership to
groups that are in a different domain. This change also aligns renames
the missing_behaviour option to lookup_failure_action to be consistent
with the other lookup DN options. THe missing_behaviour option is still
present as an alias for backwards compatibility.