Added microsoft.ad.service account module
Adds the microsoft.ad.service_account module which can be used to
manage service accounts like a group managed service account (gMSA).
It is not possible to use the microsoft.ad.object module to achieve the
same outcome as the attribute which stores what principals can retrieve
the password is a SecurityDescriptor value which is not easy to craft in
Ansible.