group - fix setting explicitly empty members
Add a fix when attempting to create a group with an explicitly empty set
of members. This fix also applies to any other add/set/remove attribute
when the resulting value is an empty value. The error only applied when
creating the AD object as the `New-AD* -OtherAttributes` parameter could
not handle an empty array value.