Filter out MSA in the LDAP inventory plugin
This change will filter out all managed service accounts in the default
LDAP filter used with the `microsoft.ad.ldap` filter plugin. This
ensures that the only objects returned as hosts are actual computer
accounts and not object classes that descend from the `computer` class
but are not actually computers.