[microsoft.ad.user] Add parameter to fail, ignore or warn if the account performing the action does not have the permissions required to modify the AD Group (#166)
This feature adds a new sub-parameter to the Groups section similar to lookup_failure_action to aid the scenario when the account used to add or remove the user from the specified AD Groups does not have appropriate permissions to perform the action.

This is achieved through wrapping the add or remove attempts around a try/catch and handling the try/catch based on the parameters specified.

Parameters accepted are fail, ignore, and warn
Default action is: fail